Makefile: add 'dist' and similar targets
This is done in a proper way, ie all the files that need to be included
in a dist tarball are explicitly listed.

In the meantime, generalize Makefile a bit.

For the sake of a better place, I take a version field from the .spec file,
feel free to change that.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Hacked-by: Konstantin Khlebnikov <koct9i@gmail.com>