Makefile: make sure git tag == src ver
It's a common situation when git tag was updated, but a version in
the source code was not (or vice versa), and it's easy to catch it.

This is exactly what the code does.

Note we can still build the binary, but can no longer build rpm or
pack a source or binary tarball.

Alternatively, we can just send -DVERSION from the command line, but
it won't work in case .git is not present, so we just stop the build
(and let the human sort it out).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>